Relocate wifi settings
